+++ src/services/answer-refiner/index.ts
@@ -6,9 +6,10 @@ import type {
   RefinementResult,
   WriteResult,
 } from "../../types";
-import { GoogleSheetsService, GoogleSheetsServiceFactory } from "../index";
 
-import { GeminiClient } from "../gemini-client";
+import { GeminiClient } from "@/services/gemini-client";
+import { GoogleSheetsService } from "@/services/google-sheets";
+import { GoogleSheetsServiceFactory } from "@/services/index";
 
 export class AnswerRefiner {
   private geminiClient: GeminiClient;
